requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
cmake==3.27.9
comm==0.2.0
connected-components-3d==3.12.4
contourpy==1.2.0
cycler==0.12.1
datasets==2.16.0
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
efficientnet-pytorch==0.7.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.46.0
frozenlist==1.4.1
fsspec==2023.10.0
helper==2.5.0
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.0
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lazy_loader==0.3
lit==17.0.6
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qudida==0.0.4
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
segmentation-models-pytorch==0.3.3
six==1.16.0
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.9.2
tokenizers==0.15.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.36.2
triton==2.0.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vtk==9.3.0
wcwidth==0.2.12
xxhash==3.4.1
yarl==1.9.4
